Add files via upload
A fully usable todolist using django , python html and css . The application consists of a login page and a register page, any user who is not authenticated gets automatically redirected to the login page as well as to the register page.  You have the confirmation of complete tasks as well as the incompleted task . On the top you have the number of pending tasks as well as incompleted tasks. User can logout easily at any time